+++ tests/test_metadataset_v2.py
@@ -30,7 +30,7 @@ from megatron.energon import (
 )
 from megatron.energon.edataclass import edataclass
 from megatron.energon.epathlib.epath import EPath
-from megatron.energon.flavors.webdataset.config import MAIN_FOLDER_NAME
+from megatron.energon.flavors.common.manifest.paths import MAIN_FOLDER_NAME
 from megatron.energon.metadataset.loader import prepare_metadataset, traverse_metadataset
 from megatron.energon.metadataset.loader_interface import DatasetBlendMode, LoadedDataset
 from megatron.energon.task_encoder.base import DefaultTaskEncoder, PackingBufferSize, stateless
